backend/pyproject.toml
[project]
name = "crimemill"
version = "0.1.0"
description = "Automated YouTube crime documentary pipeline"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "psycopg[binary,pool]>=3.2.0",
    "httpx>=0.28.0",
    "anthropic>=0.42.0",
    "boto3>=1.35.0",
    "Pillow>=11.0.0",
    "numpy>=1.26.0",
    "pydub>=0.25.1",
    "python-multipart>=0.0.18",
    "orjson>=3.10.0",
    "structlog>=24.4.0",
    "gdeltdoc>=1.5.0",
    "rapidfuzz>=3.10.0",
    "click>=8.1.0",
    "rich>=13.9.0",
    "scipy>=1.14.0",
    "pymupdf>=1.24.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.34.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "coverage>=7.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# Pydantic models need runtime access to uuid, datetime, Decimal for validation.
# FastAPI routes need runtime access to Annotated[..., Depends(...)] types.
"src/models/*.py" = ["TC001", "TC002", "TC003"]
"src/api/*.py" = ["TC001", "TC002", "TC003", "B008"]
"tests/*.py" = ["TC001", "TC002", "TC003"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring real API calls (deselect with '-m not integration')",
    "slow: marks tests that take >10 seconds",
]

[tool.coverage.run]
source = ["src"]
omit = ["src/cli.py"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 50
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ ==",
    "@overload",
]

[tool.coverage.xml]
output = "coverage.xml"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
